Added scripts for repair and preprocessing, updated bidsification

+++ climbprep/bidsify.py
@@ -6,8 +6,7 @@ from tempfile import TemporaryDirectory
 import pandas as pd
 import argparse
 
-BIDS_PATH = os.path.normpath(os.path.join('/', 'juice2', 'scr2', 'nlp', 'climblab', 'BIDS'))
-DEFAULT_TASK = 'UnknownTask'
+from climbprep.constants import *
 
 if __name__ == '__main__':
     argparser = argparse.ArgumentParser(('Convenience utility to convert fMRI source data to BIDS. '
@@ -23,20 +22,26 @@ if __name__ == '__main__':
                                          'modeling. If the task for a run is missing from the table '
                                          'or the table is missing altogether, the taskname '
                                          'will be `UnknownTask`.'))
-    argparser.add_argument('project', help='Name of BIDS project (e.g., "climblab", "evlab", etc.)')
     argparser.add_argument('subject', help="BIDS subject ID.")
-    argparser.add_argument('session', help="BIDS session ID.")
+    argparser.add_argument('-p', '--project', default='climblab', help=('Name of BIDS project (e.g., "climblab", "evlab", '
+                                                                        'etc.). Default: "climblab"'))
+    argparser.add_argument('-s', '--session', default=None, help="BIDS session ID. If ``None``, assume single-session.")
     argparser.add_argument('-t', '--tmp_dir', default=None, help=('Temporary directory to use for dcm2bids. '
                                                     'Created automatically if left blank. '
                                                     'This is useful for saving the outputs of '
-                                                    'dcm2niix if you are debugging things.')
-    argparser.add_argument('-c', '--clobber', action='store_true', help="Overwrite existing BIDS data.")
+                                                    'dcm2niix if you are debugging things.'))
+    argparser.add_argument('-c', '--clobber', action='store_true', help="Overwrite existing outputs.")
+    argparser.add_argument('-C', '--config_path', default=None, help="Path to config (json) for dcm2bids. Generated if unused.")
     args = argparser.parse_args()
 
     project_path = os.path.join(BIDS_PATH, args.project)
-    src_path = os.path.join(project_path, 'sourcedata', 'sub-%s' % args.subject, 'ses-%s' % args.session)
+    subdir = 'sub-%s' % args.subject
+    if args.session:
+        subdir = os.path.join(subdir, 'ses-%s' % args.session)
+    src_path = os.path.join(project_path, 'sourcedata', subdir)
+    assert os.path.exists(src_path), 'Path not found: %s' % src_path
     run_table_path = os.path.join(src_path, 'runs.csv')
-    out_path = os.path.join(project_path, 'sub-%s' % args.subject, 'ses-%s' % args.session)
+    out_path = os.path.join(project_path, subdir)
 
     if os.path.exists(run_table_path):
         run_table = pd.read_csv(run_table_path)
@@ -60,10 +65,13 @@ if __name__ == '__main__':
             series_number = int(jsonpath.split('_')[0])
             with open(os.path.join(tmp_dir, 'tmp_dcm2bids', 'helper', jsonpath), 'r') as f:
                 meta = json.load(f)
+            if 'BidsGuess' not in meta:
+                continue
             dtype, suffix = meta['BidsGuess']
             suffix = suffix.split('_')[-1]
             if dtype != 'discard':
                 if dtype == 'func' and len(descriptions) and \
+                        'AcquisitionTime' in meta and 'AcquisitionTime' in descriptions[-1]['criteria'] and \
                         meta['AcquisitionTime'] == descriptions[-1]['criteria']['AcquisitionTime']:
                     continue
                 description = dict(
@@ -71,10 +79,11 @@ if __name__ == '__main__':
                     suffix=suffix,
                     criteria=dict(
                         SeriesNumber=meta['SeriesNumber'],
-                        AcquisitionTime=meta['AcquisitionTime'],
                         SeriesDescription=meta['SeriesDescription']
                     )
                 )
+                if 'AcquisitionTime' in meta:
+                    description['criteria']['AcquisitionTime'] = meta['AcquisitionTime']
                 if dtype == 'func':
                     task_name = run_map.get(series_number, {}).get('TaskName', DEFAULT_TASK)
                     description['custom_entities'] = 'task-%s' % task_name
@@ -86,29 +95,41 @@ if __name__ == '__main__':
 
         descriptions = dict(descriptions=descriptions)
 
-        config_path = os.path.join(src_path, 'dcm2bids_config.json')
-        with open(config_path, 'w') as f:
-            json.dump(descriptions, f, indent=2)
+        if args.config_path:
+            config_path = args.config_path
+        else:
+            config_path = os.path.join(src_path, 'dcm2bids_config.json')
+            with open(config_path, 'w') as f:
+                json.dump(descriptions, f, indent=2)
 
         if args.clobber:
             clobber = ' --clobber'
         else:
             clobber = ''
-        os.system('dcm2bids -d %s -p %s -s %s -c %s -o %s --auto_extract_entities%s --skip_dcm2niix --force_dcm2bids' % (
-            os.path.join(tmp_dir, 'tmp_dcm2bids', 'helper'), args.subject, args.session, config_path, project_path, clobber))
+        if args.session:
+            session = ' -s %s' % args.session
+        else:
+            session = ''
+        os.system('dcm2bids -d %s -p %s%s -c %s -o %s --auto_extract_entities%s --skip_dcm2niix --force_dcm2bids' % (
+            os.path.join(tmp_dir, 'tmp_dcm2bids', 'helper'), args.subject, session, config_path, project_path, clobber))
 
-        for path in os.listdir(os.path.join(out_path, 'func')):
-            if path.endswith('.json'):
-                filepath = os.path.join(out_path, 'func', path)
-                with open(filepath, 'r') as f:
-                    sidecar = json.load(f)
-                if 'EventsFile' in sidecar:
-                    events_path_new = os.path.join(out_path, 'func', path.replace('bold.json', 'events.tsv'))
-                    if events_path_new != sidecar['EventsFile']:
-                        shutil.copy2(sidecar['EventsFile'], events_path_new)
-                        sidecar['EventsFile'] = events_path_new
-                with open(filepath, 'w') as f:
-                    json.dump(sidecar, f, indent=2)
+        dirnames = ('func', 'derived')
+        for dirname in dirnames:
+            dir_path = os.path.join(out_path, dirname)
+            if not os.path.exists(dir_path):
+                continue
+            for path in os.listdir(dir_path):
+                if path.endswith('.json'):
+                    filepath = os.path.join(dir_path, path)
+                    with open(filepath, 'r') as f:
+                        sidecar = json.load(f)
+                    if 'EventsFile' in sidecar:
+                        events_path_new = os.path.join(dir_path, path.replace('bold.json', 'events.tsv'))
+                        if events_path_new != sidecar['EventsFile']:
+                            shutil.copy2(sidecar['EventsFile'], events_path_new)
+                            sidecar['EventsFile'] = events_path_new
+                    with open(filepath, 'w') as f:
+                        json.dump(sidecar, f, indent=2)
 
         os.system('bids-validator %s' % project_path)
